# artifacts/api-server/src/routes/preferences.ts
import { Router, type IRouter } from "express";
import { getAuth } from "@clerk/express";
import { eq } from "drizzle-orm";
import {
  db,
  userPreferences,
  commitments,
  schedules,
  assignments,
} from "@workspace/db";
import { resolveOwnerId } from "../lib/auth";
import {
  GetPreferencesQueryParams,
  UpdatePreferencesBody,
  ClaimGuestDataBody,
} from "@workspace/api-zod";

const router: IRouter = Router();

function serializePreferences(
  row: typeof userPreferences.$inferSelect | undefined,
) {
  return {
    wakeTime: row?.wakeTime ?? null,
    bedTime: row?.bedTime ?? null,
    mealTimes: row?.mealTimes ?? null,
    notes: row?.notes ?? null,
  };
}

router.get("/preferences", async (req, res) => {
  const { deviceId } = GetPreferencesQueryParams.parse(req.query);
  const ownerId = resolveOwnerId(req, deviceId);
  const [row] = await db
    .select()
    .from(userPreferences)
    .where(eq(userPreferences.ownerId, ownerId));
  res.json(serializePreferences(row));
});

router.put("/preferences", async (req, res) => {
  const body = UpdatePreferencesBody.parse(req.body);
  const ownerId = resolveOwnerId(req, body.deviceId);

  const values = {
    wakeTime: body.wakeTime ?? null,
    bedTime: body.bedTime ?? null,
    mealTimes: body.mealTimes ?? null,
    notes: body.notes ?? null,
  };

  const [row] = await db
    .insert(userPreferences)
    .values({ ownerId, ...values })
    .onConflictDoUpdate({
      target: userPreferences.ownerId,
      set: { ...values, updatedAt: new Date() },
    })
    .returning();

  res.json(serializePreferences(row));
});

// Only ids generated by the web guest flow may be claimed. Web guest ids are
// always "guest-<uuidv4>" (see study-flow-web's use-device-id.ts), where the
// UUID is 122 bits of browser-generated randomness — possession of the id is
// the proof of ownership, like a bearer token. The strict format check means
// a signed-in caller can never "claim" (steal) rows belonging to a Clerk
// account ("user_...") or to a mobile install (raw UUID device ids).
const WEB_GUEST_ID_PATTERN =
  /^guest-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;

// Cheap in-memory rate limit so a signed-in user can't brute-force guest ids.
const claimAttempts = new Map<string, { count: number; resetAt: number }>();
const CLAIM_LIMIT = 10;
const CLAIM_WINDOW_MS = 60 * 60 * 1000;

function isClaimRateLimited(userId: string): boolean {
  const now = Date.now();
  const entry = claimAttempts.get(userId);
  if (!entry || now >= entry.resetAt) {
    claimAttempts.set(userId, { count: 1, resetAt: now + CLAIM_WINDOW_MS });
    return false;
  }
  entry.count += 1;
  return entry.count > CLAIM_LIMIT;
}

router.post("/claim-guest-data", async (req, res) => {
  const auth = getAuth(req);
  if (!auth?.userId) {
    res.status(401).json({ message: "Sign in required to claim guest data" });
    return;
  }
  const body = ClaimGuestDataBody.parse(req.body);
  const guestId = body.guestDeviceId;

  if (!guestId || !WEB_GUEST_ID_PATTERN.test(guestId)) {
    res.status(400).json({ message: "Invalid guest device id" });
    return;
  }

  if (isClaimRateLimited(auth.userId)) {
    res
      .status(429)
      .json({ message: "Too many claim attempts. Try again later." });
    return;
  }

  const userId = auth.userId;
  let claimedCount = 0;

  const movedCommitments = await db
    .update(commitments)
    .set({ deviceId: userId })
    .where(eq(commitments.deviceId, guestId))
    .returning({ id: commitments.id });
  claimedCount += movedCommitments.length;

  const movedSchedules = await db
    .update(schedules)
    .set({ deviceId: userId })
    .where(eq(schedules.deviceId, guestId))
    .returning({ id: schedules.id });
  claimedCount += movedSchedules.length;

  const movedAssignments = await db
    .update(assignments)
    .set({ deviceId: userId })
    .where(eq(assignments.deviceId, guestId))
    .returning({ id: assignments.id });
  claimedCount += movedAssignments.length;

  // Preferences: only move the guest row if the user doesn't already have one.
  const [existingUserPrefs] = await db
    .select()
    .from(userPreferences)
    .where(eq(userPreferences.ownerId, userId));
  const [guestPrefs] = await db
    .select()
    .from(userPreferences)
    .where(eq(userPreferences.ownerId, guestId));
  if (guestPrefs) {
    if (existingUserPrefs) {
      await db
        .delete(userPreferences)
        .where(eq(userPreferences.ownerId, guestId));
    } else {
      await db
        .update(userPreferences)
        .set({ ownerId: userId, updatedAt: new Date() })
        .where(eq(userPreferences.ownerId, guestId));
      claimedCount += 1;
    }
  }

  res.json({ claimedCount });
});

export default router;