# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CommitmentType = typeof CommitmentType[keyof typeof CommitmentType];


export const CommitmentType = {
  class: 'class',
  extracurricular: 'extracurricular',
  routine: 'routine',
} as const;

export type DayOfWeek = typeof DayOfWeek[keyof typeof DayOfWeek];


export const DayOfWeek = {
  mon: 'mon',
  tue: 'tue',
  wed: 'wed',
  thu: 'thu',
  fri: 'fri',
  sat: 'sat',
  sun: 'sun',
} as const;

export interface Commitment {
  id: string;
  deviceId: string;
  title: string;
  type: CommitmentType;
  daysOfWeek: DayOfWeek[];
  /** 24h time, "HH:mm" */
  startTime: string;
  /** 24h time, "HH:mm" */
  endTime: string;
  notes?: string | null;
  createdAt: string;
}

export interface CreateCommitmentInput {
  deviceId: string;
  title: string;
  type: CommitmentType;
  daysOfWeek: DayOfWeek[];
  startTime: string;
  endTime: string;
  notes?: string | null;
}

export interface UpdateCommitmentInput {
  deviceId: string;
  title?: string;
  type?: CommitmentType;
  daysOfWeek?: DayOfWeek[];
  startTime?: string;
  endTime?: string;
  notes?: string | null;
}

export interface ExtractCommitmentsInput {
  deviceId: string;
  /** The schedule photo, as a data URL (e.g. "data:image/png;base64,...") or a bare base64-encoded image string. */
  imageBase64: string;
}

export interface ExtractCommitmentsTextInput {
  deviceId: string;
  /** Free-text description of the student's weekly schedule, e.g. "School 8am-3pm weekdays, soccer 4-5:30 Tue/Thu". */
  description: string;
}

export interface UserPreferences {
  /** When the student usually wakes up (e.g. "7:00 AM"). */
  wakeTime?: string | null;
  /** When the student usually goes to bed (e.g. "10:30 PM"). */
  bedTime?: string | null;
  /** Usual meal times, free text (e.g. "breakfast 7:30, lunch 12, dinner 6:30"). */
  mealTimes?: string | null;
  /** Anything else the AI should keep in mind when scheduling. */
  notes?: string | null;
}

export interface UpdatePreferencesInput {
  deviceId: string;
  wakeTime?: string | null;
  bedTime?: string | null;
  mealTimes?: string | null;
  notes?: string | null;
}

export interface ClaimGuestDataInput {
  /** The anonymous device id whose data should be claimed. */
  guestDeviceId: string;
}

export interface ClaimGuestDataResult {
  /** Total number of rows reassigned to the signed-in user. */
  claimedCount: number;
}

/**
 * An assignment or task the student needs to work on.
 */
export interface Task {
  title: string;
  /** When it is due, free text (e.g. "tomorrow", "Friday", "2026-07-10"). */
  dueDate: string;
  /** Rough estimate of how long the task will take, in minutes. */
  estimatedMinutes?: number | null;
  notes?: string | null;
}

export type ScheduleBlockCategory = typeof ScheduleBlockCategory[keyof typeof ScheduleBlockCategory];


export const ScheduleBlockCategory = {
  class: 'class',
  extracurricular: 'extracurricular',
  routine: 'routine',
  homework: 'homework',
  break: 'break',
  free: 'free',
} as const;

export interface ScheduleBlock {
  id: string;
  day: DayOfWeek;
  startTime: string;
  endTime: string;
  title: string;
  category: ScheduleBlockCategory;
  notes?: string | null;
}

export type ScheduleStatus = typeof ScheduleStatus[keyof typeof ScheduleStatus];


export const ScheduleStatus = {
  needs_clarification: 'needs_clarification',
  complete: 'complete',
} as const;

export type ScheduleScope = typeof ScheduleScope[keyof typeof ScheduleScope];


export const ScheduleScope = {
  day: 'day',
  week: 'week',
} as const;

export interface ScheduleSummary {
  id: string;
  deviceId: string;
  scope: ScheduleScope;
  status: ScheduleStatus;
  createdAt: string;
}

export interface Schedule {
  id: string;
  deviceId: string;
  scope: ScheduleScope;
  status: ScheduleStatus;
  blocks: ScheduleBlock[];
  clarifyingQuestions: string[];
  createdAt: string;
}

export interface ScheduleBlockInput {
  /** Stable block id. Include for existing blocks so Google Calendar sync mapping is preserved; omit for newly added blocks and the server will assign one. */
  id?: string;
  day: DayOfWeek;
  startTime: string;
  endTime: string;
  title: string;
  category: ScheduleBlockCategory;
  notes?: string | null;
}

export interface UpdateScheduleInput {
  deviceId: string;
  blocks: ScheduleBlockInput[];
}

export interface ReviseScheduleInput {
  deviceId: string;
  /** Natural-language description of the change the student wants. */
  instruction: string;
}

export interface ClarificationAnswer {
  question: string;
  answer: string;
}

export interface GenerateScheduleInput {
  deviceId: string;
  scope: ScheduleScope;
  /** Pass the id returned from a prior needs_clarification response to continue that draft */
  draftId?: string | null;
  answers?: ClarificationAnswer[];
  /** Assignments/tasks the student needs time for in this schedule. */
  tasks?: Task[];
  /** The student's scheduling preferences (sleep, meals, extra notes). When provided, the AI should use these instead of asking clarifying questions. */
  preferences?: UserPreferences;
}

export interface GoogleCalendarStatus {
  connected: boolean;
}

export interface SyncScheduleGoogleCalendarInput {
  /** IANA timezone (e.g. "America/New_York") used to place each block's start/end times on the student's Google Calendar. */
  timeZone: string;
}

export interface SyncScheduleGoogleCalendarResult {
  syncedCount: number;
  calendarUrl: string;
}

export interface ScheduleCalendarSync {
  blockId: string;
  googleEventId: string;
  createdAt: string;
}

export interface GenerateScheduleResult {
  id: string;
  status: ScheduleStatus;
  questions?: string[];
  schedule?: Schedule;
}

export type AssignmentSource = typeof AssignmentSource[keyof typeof AssignmentSource];


export const AssignmentSource = {
  canvas: 'canvas',
  classroom: 'classroom',
} as const;

export interface Assignment {
  id: string;
  deviceId: string;
  source: AssignmentSource;
  externalId: string;
  courseName?: string | null;
  title: string;
  /** ISO date or datetime when the assignment is due. */
  dueDate: string;
  /** Link to the assignment in the source system. */
  url?: string | null;
  createdAt: string;
}

export interface IntegrationsStatus {
  canvasConnected: boolean;
  canvasBaseUrl?: string | null;
  /** Whether the signed-in user's Google account is connected (Classroom access is granted through the same connection). */
  classroomConnected: boolean;
}

export interface ConnectCanvasInput {
  deviceId: string;
  /** The school's Canvas URL, e.g. "https://myschool.instructure.com". */
  baseUrl: string;
  /** Canvas personal access token generated in Account → Settings → New Access Token. */
  accessToken: string;
}

export interface ImportAssignmentsInput {
  deviceId: string;
}

export interface ImportAssignmentsResult {
  /** Number of new assignments imported (existing ones are updated, not duplicated). */
  importedCount: number;
  assignments: Assignment[];
}

export type ListCommitmentsParams = {
deviceId: string;
};

export type GetPreferencesParams = {
deviceId: string;
};

export type DeleteCommitmentParams = {
deviceId: string;
};

export type ListSchedulesParams = {
deviceId: string;
};

export type GetScheduleParams = {
deviceId: string;
};

export type DeleteScheduleParams = {
deviceId: string;
};

export type GetIntegrationsStatusParams = {
deviceId: string;
};

export type DisconnectCanvasParams = {
deviceId: string;
};

export type ListAssignmentsParams = {
deviceId: string;
};

export type DeleteAssignmentParams = {
deviceId: string;
};