# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /commitments:
    get:
      operationId: listCommitments
      tags: [commitments]
      summary: List commitments for a device
      parameters:
        - name: deviceId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of commitments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commitment"
    post:
      operationId: createCommitment
      tags: [commitments]
      summary: Create a commitment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommitmentInput"
      responses:
        "201":
          description: Created commitment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commitment"
  /commitments/extract-image:
    post:
      operationId: extractCommitmentsFromImage
      tags: [commitments]
      summary: Extract recurring commitments from a photo of a schedule/timetable
      description: >
        Accepts a photo of a student's timetable or schedule and uses AI vision
        to extract recurring commitments (classes, activities, routines). The
        extracted commitments are saved for the device and returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractCommitmentsInput"
      responses:
        "201":
          description: Extracted and saved commitments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commitment"
  /commitments/extract-text:
    post:
      operationId: extractCommitmentsFromText
      tags: [commitments]
      summary: Extract recurring commitments from a plain-text description
      description: >
        Accepts a free-text description of a student's weekly schedule
        (e.g. "I have school 8-3, soccer practice 4-5 on Tuesdays") and uses
        AI to extract recurring commitments. The extracted commitments are
        saved for the device and returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractCommitmentsTextInput"
      responses:
        "201":
          description: Extracted and saved commitments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commitment"
  /preferences:
    get:
      operationId: getPreferences
      tags: [preferences]
      summary: Get the saved scheduling preferences for a user/device
      parameters:
        - name: deviceId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Saved preferences (empty values if never saved)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
    put:
      operationId: updatePreferences
      tags: [preferences]
      summary: Save scheduling preferences for a user/device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePreferencesInput"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
  /claim-guest-data:
    post:
      operationId: claimGuestData
      tags: [preferences]
      summary: Reassign data created as a signed-out guest to the signed-in user
      description: >
        After a guest signs in, transfers commitments, schedules, assignments
        and preferences owned by the anonymous guest device id to the
        signed-in user's account. Requires an authenticated session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimGuestDataInput"
      responses:
        "200":
          description: Claim result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimGuestDataResult"
  /commitments/{id}:
    patch:
      operationId: updateCommitment
      tags: [commitments]
      summary: Update a commitment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommitmentInput"
      responses:
        "200":
          description: Updated commitment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commitment"
    delete:
      operationId: deleteCommitment
      tags: [commitments]
      summary: Delete a commitment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: deviceId
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    CommitmentType:
      type: string
      enum: [class, extracurricular, routine]
    DayOfWeek:
      type: string
      enum: [mon, tue, wed, thu, fri, sat, sun]
    Commitment:
      type: object
      properties:
        id:
          type: string
        deviceId:
          type: string
        title:
          type: string
        type:
          $ref: "#/components/schemas/CommitmentType"
        daysOfWeek:
          type: array
          items:
            $ref: "#/components/schemas/DayOfWeek"
        startTime:
          type: string
          description: 24h time, "HH:mm"
        endTime:
          type: string
          description: 24h time, "HH:mm"
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - deviceId
        - title
        - type
        - daysOfWeek
        - startTime
        - endTime
        - createdAt
    CreateCommitmentInput:
      type: object
      properties:
        deviceId:
          type: string
        title:
          type: string
        type:
          $ref: "#/components/schemas/CommitmentType"
        daysOfWeek:
          type: array
          items:
            $ref: "#/components/schemas/DayOfWeek"
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - deviceId
        - title
        - type
        - daysOfWeek
        - startTime
        - endTime
    UpdateCommitmentInput:
      type: object
      properties:
        deviceId:
          type: string
        title:
          type: string
        type:
          $ref: "#/components/schemas/CommitmentType"
        daysOfWeek:
          type: array
          items:
            $ref: "#/components/schemas/DayOfWeek"
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - deviceId
    ExtractCommitmentsInput:
      type: object
      properties:
        deviceId:
          type: string
        imageBase64:
          type: string
          description: >
            The schedule photo, as a data URL (e.g. "data:image/png;base64,...")
            or a bare base64-encoded image string.
      required:
        - deviceId
        - imageBase64
    ExtractCommitmentsTextInput:
      type: object
      properties:
        deviceId:
          type: string
        description:
          type: string
          description: >
            Free-text description of the student's weekly schedule, e.g.
            "School 8am-3pm weekdays, soccer 4-5:30 Tue/Thu".
      required:
        - deviceId
        - description
    UserPreferences:
      type: object
      properties:
        wakeTime:
          type: string
          nullable: true
          description: 'When the student usually wakes up (e.g. "7:00 AM").'
        bedTime:
          type: string
          nullable: true
          description: 'When the student usually goes to bed (e.g. "10:30 PM").'
        mealTimes:
          type: string
          nullable: true
          description: 'Usual meal times, free text (e.g. "breakfast 7:30, lunch 12, dinner 6:30").'
        notes:
          type: string
          nullable: true
          description: Anything else the AI should keep in mind when scheduling.
      required: []
    UpdatePreferencesInput:
      type: object
      properties:
        deviceId:
          type: string
        wakeTime:
          type: string
          nullable: true
        bedTime:
          type: string
          nullable: true
        mealTimes:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - deviceId
    ClaimGuestDataInput:
      type: object
      properties:
        guestDeviceId:
          type: string
          description: The anonymous device id whose data should be claimed.
      required:
        - guestDeviceId
    ClaimGuestDataResult:
      type: object
      properties:
        claimedCount:
          type: integer
          description: Total number of rows reassigned to the signed-in user.
      required:
        - claimedCount
...