# lib/api-zod/src/generated/types/claimGuestDataInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClaimGuestDataInput {
  /** The anonymous device id whose data should be claimed. */
  guestDeviceId: string;
}