# lib/api-zod/src/generated/types/claimGuestDataResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClaimGuestDataResult {
  /** Total number of rows reassigned to the signed-in user. */
  claimedCount: number;
}