# lib/api-zod/src/generated/types/extractCommitmentsTextInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractCommitmentsTextInput {
  deviceId: string;
  /** Free-text description of the student's weekly schedule, e.g. "School 8am-3pm weekdays, soccer 4-5:30 Tue/Thu". */
  description: string;
}