# lib/api-zod/src/generated/types/updatePreferencesInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePreferencesInput {
  deviceId: string;
  wakeTime?: string | null;
  bedTime?: string | null;
  mealTimes?: string | null;
  notes?: string | null;
}