# lib/api-zod/src/generated/types/userPreferences.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserPreferences {
  /** When the student usually wakes up (e.g. "7:00 AM"). */
  wakeTime?: string | null;
  /** When the student usually goes to bed (e.g. "10:30 PM"). */
  bedTime?: string | null;
  /** Usual meal times, free text (e.g. "breakfast 7:30, lunch 12, dinner 6:30"). */
  mealTimes?: string | null;
  /** Anything else the AI should keep in mind when scheduling. */
  notes?: string | null;
}